+++ lib/shrine/plugins/included.rb
@@ -7,13 +7,14 @@ class Shrine
     # [doc/plugins/included.md]: https://github.com/shrinerb/shrine/blob/master/doc/plugins/included.md
     module Included
       def self.configure(uploader, &block)
-        uploader.opts[:included_block] = block
+        uploader.opts[:included] ||= {}
+        uploader.opts[:included][:block] = block
       end
 
       module AttachmentMethods
         def included(klass)
           super
-          klass.instance_exec(@name, &shrine_class.opts[:included_block])
+          klass.instance_exec(@name, &shrine_class.opts[:included][:block])
         end
       end
     end